+++ src/material/ImageTexture.h
@@ -31,7 +31,7 @@ public:
 
     void prepareArray(
         ArrayTexture& arr,
-        uint32_t layer) override;
+        uint16_t layer) override;
 
     void preparePlain();
     void prepareKtx();
